refactor: 코드 리펙토링

+++ src/main/java/geunbon/geunboni/domain/auth/service/AuthServiceImpl.java
@@ -11,11 +11,13 @@ import geunbon.geunboni.domain.user.entity.User;
 import geunbon.geunboni.domain.user.repository.UserRepository;
 import geunbon.geunboni.global.error.CustomException;
 import geunbon.geunboni.global.error.ErrorCode;
-import geunbon.geunboni.global.security.details.CustomUserDetails;
 import geunbon.geunboni.global.security.jwt.dto.Jwt;
 import geunbon.geunboni.global.security.jwt.enums.JwtType;
 import geunbon.geunboni.global.security.jwt.provider.JwtProvider;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -90,22 +92,36 @@ public class AuthServiceImpl implements AuthService {
 
     @Transactional
     @Override
-    public void deleteAccount(CustomUserDetails customUserDetails) {
-        User user = customUserDetails.user();
+    public void deleteAccount() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication.getPrincipal() instanceof UserDetails userDetails) {
+            String email = userDetails.getUsername();
 
-        userRepository.delete(user);
+            User user = userRepository.findByEmail(email).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+
+            userRepository.delete(user);
+        } else {
+            throw new CustomException(ErrorCode.USER_NOT_FOUND);
+        }
     }
 
     @Transactional
     @Override
-    public void changePassword(ChangePasswordRequest request, CustomUserDetails customUserDetails) {
-        User user = customUserDetails.user();
+    public void changePassword(ChangePasswordRequest request) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication.getPrincipal() instanceof UserDetails userDetails) {
+            String email = userDetails.getUsername();
 
-        if (!passwordEncoder.matches(request.getCurrentPassword(), user.getPassword())) {
-            throw new CustomException(ErrorCode.WRONG_PASSWORD);
+            User user = userRepository.findByEmail(email).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+
+            if (!passwordEncoder.matches(request.getCurrentPassword(), user.getPassword())) {
+                throw new CustomException(ErrorCode.WRONG_PASSWORD);
+            }
+
+            user.setPassword(passwordEncoder.encode(request.getNewPassword()));
+            userRepository.save(user);
+        } else {
+            throw new CustomException(ErrorCode.USER_NOT_FOUND);
         }
-
-        user.setPassword(passwordEncoder.encode(request.getNewPassword()));
-        userRepository.save(user);
     }
 }
